requirements.txt
PyQt5
PySerial
PyInstaller